+++ fabric/src/main/java/io/github/sakurawald/module/initializer/kit/command/argument/adapter/KitNameArgumentTypeAdapter.java
@@ -7,8 +7,8 @@ import com.mojang.brigadier.context.CommandContext;
 import io.github.sakurawald.core.auxiliary.minecraft.CommandHelper;
 import io.github.sakurawald.core.command.argument.adapter.abst.BaseArgumentTypeAdapter;
 import io.github.sakurawald.core.command.argument.structure.Argument;
-import io.github.sakurawald.module.initializer.kit.KitInitializer;
 import io.github.sakurawald.module.initializer.kit.command.argument.wrapper.KitName;
+import io.github.sakurawald.module.initializer.kit.service.KitService;
 import net.minecraft.server.command.ServerCommandSource;
 
 import java.util.List;
@@ -37,6 +37,6 @@ public class KitNameArgumentTypeAdapter extends BaseArgumentTypeAdapter {
 
     @Override
     public RequiredArgumentBuilder<ServerCommandSource, ?> makeRequiredArgumentBuilder(String argumentName) {
-        return super.makeRequiredArgumentBuilder(argumentName).suggests(CommandHelper.Suggestion.iterable(KitInitializer::listKitNames));
+        return super.makeRequiredArgumentBuilder(argumentName).suggests(CommandHelper.Suggestion.iterable(KitService::listKitNames));
     }
 }