+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 - [head] fix: should not eat the cursor stack when the price if `FREE`.
 - [head] fix: should decrease the amount of heads on right click when the price is `FREE`.
+- (> MC 1.20.2)[kit & deathlog] fix: failed to save NBT file.
 - [pvp] refactor: replace the language key `pvp.status` with two new language key `pvp.status.on` and `pvp.status.off`. (Should not use composition for language keys, it's complex.)